refactored analytics collection into analyticsRecorder (AIP 23 implementation)
